# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des utilisateurs et vendeurs.",
        "title": "o'z api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support API",
            "email": "goliyaohugues@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère toutes les commandes de l'utilisateur connecté avec leurs items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Liste toutes les commandes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Utilisateur non identifié",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crée une commande, décrémente le stock et enregistre les mouvements SALE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Créer une commande",
                "parameters": [
                    {
                        "description": "Données de la commande",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Données invalides ou stock insuffisant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère une commande par son UUID avec tous ses items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Détails d'une commande",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la commande",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Format d'identifiant invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Commande introuvable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Annule une commande et restaure le stock de chaque item",
                "tags": [
                    "Orders"
                ],
                "summary": "Annuler une commande",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de la commande",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Commande introuvable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Order": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItem"
                    }
                },
                "orderUUID": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "shop_id": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.OrderItem": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "unit_price": {
                    "type": "integer"
                }
            }
        },
...
        "transport.createOrderRequest": {
            "type": "object",
            "required": [
                "items",
                "shop_id"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/transport.orderItemRequest"
                    }
                },
                "shop_id": {
                    "type": "integer"
                }
            }
        },
...
        "transport.orderItemRequest": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" suivi d'un espace et du token JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}